% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longAlphaMarginal.R
\name{longAlphaMarginal}
\alias{longAlphaMarginal}
\title{Marginal correlation estimation for X->M}
\usage{
longAlphaMarginal(X, M, COV, wave)
}
\arguments{
\item{X}{repeated measured exposure / treatment / input.}

\item{M}{time-invariant repeated measure of M high-dimensional mediators.}

\item{COV}{covariants, can be null.}

\item{wave}{is used to indicate the measurement occasion in repeated measures,
representing which time point a sample was measured.}
}
\value{
\code{Est_alpha}: a matrix of the main results of X->M margin.
First row is the estimated coefficients,
Second row is the estimated variance.
Third row is the estimated p-value.
}
\description{
Marginal correlation estimation for X->M
}
