% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longBetaMarginal.R
\name{longBetaMarginal}
\alias{longBetaMarginal}
\title{Marginal correlation estimation for M->Y}
\usage{
longBetaMarginal(Y, X, M, COV, id, wave, method = c("lmm", "gee"))
}
\arguments{
\item{Y}{repeated measured outcomes for n individuals.}

\item{X}{repeated measured exposure / treatment / input.}

\item{M}{time-invariant repeated measure of M high-dimensional mediators.}

\item{COV}{covariants, can be null.}

\item{id}{the id of all the individuals.}

\item{wave}{is used to indicate the measurement occasion in repeated measures,
representing which time point a sample was measured.}

\item{method}{method used in this framework. Choices include "lmm" (linear mixed effect model),
"gee" (generalized estimating equation).}
}
\value{
\code{Est_beta}: a matrix of the main results of M->Y margin.
First row is the estimated coefficients,
Second row is the estimated variance.
Third row is the estimated p-value.
}
\description{
Marginal correlation estimation for M->Y
}
