% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMediation.R
\name{longMediation}
\alias{longMediation}
\title{MAIN FUNCTION}
\usage{
longMediation(
  Y,
  X,
  M,
  COV = NULL,
  id,
  wave,
  topN = NULL,
  SIS_by = c("coef", "pval"),
  method = c("lmm", "gee"),
  verbose = F
)
}
\arguments{
\item{Y}{repeated measured outcomes for n individuals.}

\item{X}{repeated measured exposure / treatment / input.}

\item{M}{time-invariant repeated measure of M high-dimensional mediators.}

\item{COV}{covariants, can be null.}

\item{id}{the id of all the individuals.}

\item{wave}{is used to indicate the measurement occasion in repeated measures,
representing which time point a sample was measured.}

\item{topN}{the number of top variables selected during the SIS
(Sure Independence Screening) step. If it is null, use default topN = 2*n/log(n),
n is the total sample size. It can be either a single numeric value or a vector
of two numeric values. If it is a single value, the top topN variables are
selected in both the X->M and M->Y steps. If it is a vector of two values,
the first value (topN1) indicates the number of top variables selected in the
X->M margin, and the second value (topN2) indicates the number of top variables
selected in the M->Y margin}

\item{SIS_by}{specify the marginal correlation method for conducting the SIS.
The user can choose between two options: "pval" (sort by p-value) or
"coef" (sort by absolute value of coefficient).}

\item{method}{method used in this framework. Choices include
"lmm" (linear mixed effect model),
"gee" (generalized estimating equation).}

\item{verbose}{logical: whether to provides additional detailed information
or messages during the execution of this function (default = FALSE).}
}
\value{
A list with the following members:

\code{result}: a data frame of main results.
sob_pval_fdr (FDR-BH adjusted Sobel test p-value).
sob_pval_bon (Bonferroni adjusted Sobel test p-value).
sob_pval_by (FDR-BY adjusted Sobel test p-value).
join_pval_fdr (FDR-BH adjusted Joint significant test p-value).
join_pval_bon (Bonferroni adjusted Joint significant test p-value).
join_pval_by (FDR-BY adjusted Joint significant test p-value).
ab_coef (Estimated indirect effect, alpha * beta).
ab_var (Variance of estimated indirect effect).
conf_low (Lower confidence bound).
conf_up (Upper confidence bound).
sob_pval(Sobel test p-value before multiple hypothesis adjustment).
alpha_coef (Estimated alpha coefficient).
alpha_var (Estimated alpha variance).
alpha_pval(Estimated alpha p-value).
beta_coef(Estimated beta coefficient).
beta_var (Estimated beta variance).
beta_pval (Estimated beta  p-value).

\code{subID_SIS}: ID of the intersect SIS selected mediators.

\code{subM_SIS}: a matrix of the selected mediators in step 1.

\code{Beta.margin}: marginal correlation matrix in M->Y margin.

\code{Alpha.margin}: marginal correlation matrix in X->M margin.

\code{elapsed}: elapsed time, units = mins.
}
\description{
MAIN FUNCTION
}
\examples{
# Example data for demonstration, with 200 individuals and 500 mediators:

data(example_lmm)
n = length(unique(example_lmm$id))
topN = ceiling(2 * n/log(n))
results_lmm = longMediation(Y = example_lmm$Y, X = example_lmm$X,
M = example_lmm$M, COV = example_lmm$COV, id = example_lmm$id,
wave = example_lmm$wave, topN = topN, SIS_by = "coef",
method = "lmm", verbose = TRUE)

data(example_gee)
n = length(unique(example_gee$id))
topN = ceiling(2 * n/log(n))
results_gee = longMediation(Y = example_gee$Y, X = example_gee$X,
M = example_gee$M, COV = example_gee$COV, id = example_gee$id,
wave = example_gee$wave, topN = topN, SIS_by = "coef",
method = "gee", verbose = TRUE)
}
